Re-organised `common.tools` and `utils`
