+++ src/ipbb/cmds/ipbus.py
@@ -9,7 +9,7 @@ from rich.prompt import Confirm
 from os.path import join, split, exists, abspath, splitext, relpath, basename
 from ..console import cprint, console
 from ..defaults import kProjAreaFile, kProjUserFile
-from ..tools.common import which, DEFAULT_ENCODING
+from ..utils import which, DEFAULT_ENCODING
 from ..utils import DirSentry, formatDictTable
 from .common import addrtab
