+++ src/ipbb/generators/modelsimproject.py
@@ -2,7 +2,7 @@ import time
 import os
 
 from os.path import join, split, exists, splitext, abspath
-from ..tools.common import SmartOpen
+from ..utils import SmartOpen
 
 # --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
